created and finished dashboard components
